third_party/blink/web_tests/platform/generic/fast/events/hit-test-cache-iframes-expected.txt

Ensure hit test cache is cleared across iframes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS internals.hitTestCacheHits(document) is 0
PASS internals.hitTestCacheHits(document) is 0
PASS internals.hitTestCacheHits(document) is 0
PASS successfullyParsed is true

TEST COMPLETE
